requirements.txt
pybullet~=2.5.5-eglpatched
numpy~=1.17.4
matplotlib~=3.1.2
pandas~=0.25.3
scikit-learn~=0.22.1
setuptools~=62.3.2
tqdm~=4.41.1
colorama~=0.4.3
yaml~=0.2.5
pyyaml~=5.1
opencv-python~=4.1.2.30
pytz~=2022.1
scipy~=1.4.1
imageio~=2.6.1
glumpy~=1.1.0
pillow~=6.0.0
torchvision~=0.4.2
cosypose~=1.0.0
transforms3d~=0.3.1
pinocchio~=2.3.1
eigenpy~=2.6.0
trimesh~=3.5.16
psutil~=5.9.1
seaborn~=0.11.2
ipython~=7.33.0
bokeh~=1.4.0
plyfile~=0.7.1
wget~=3.2
simplejson~=3.17.0
torchnet~=0.0.4
joblib~=0.14.1
dask~=2022.2.0
distributed~=2.9.1
xarray~=0.14.1
pymeshlab~=2022.2.post2
zerorpc~=0.6.3
efficientpose~=1.0.0
cython~=0.29.21
six~=1.16.0